requirements.txt
pypdf
faiss-cpu
langchain-community
langchain-core
langchain-text-splitters
langchain
streamlit